0.9.1
The Superpowered Audio SDK now supports tvOS

Every Superpowered feature is available on tvOS, both natively on the
device and in the simulator. The included SuperpoweredTVExample project
shows a very simple, just a few lines long code to play with the
SuperpoweredWhoosh effect on the Apple TV, using the remote control.

The new SuperpoweredtvOSAudioIO class provides an easy wrapper for
audio output. Please note, Apple did not provide developers access to
the audio input in tvOS. Testing in the simulator may show some
warnings, as Xcode is buggy in building static libraries for tvOS. You
can safely ignore the warnings in this case.

New parameters for the SuperpoweredAdvancedAudioPlayer

Minimum and maximum time stretching limits are adjustable. You can also
set the behavior when a reverse loop reaches the loop's beginning.

Rewritten iOS audio input/output

The SuperpoweredIOSAudioOutput class is now SuperpoweredIOSAudioIO.
Audio input can be set with the appropriate audio session category.
Routing changes are handled with less delay, and asking for recording
permissions are built-in.

Android improvements

Stream type can be set for the SuperpoweredAndroidAudioIO, such as
SL_ANDROID_STREAM_MEDIA for normal playback and SL_ANDROID_STREAM_VOICE
for the speakerphone. Android project configuration is simpler now, as
the NDK path must be set in local.properties only.

Other improvements

- SuperpoweredHasNonFinite function in SuperpoweredSimple to check for
NaN or infinity samples.
- Echo sound bugfix.