Compile without microhone support
This changes enables Superpowered to be easily compiled without support for the microphone.

The reason for this change is an automatic binary rejection when submitting to the App Store.  When submitting a binary, the App Store detects calls to the requestRecordPermission API and automatically assumes the app uses the microphone, even if it actually doesn’t.  As of iOS 10, apps that use the microphone must contain an NSMicrophoneUsageDescription key with a string value explaining to the user how the app uses this data.  This is problematic for apps that don’t actually use the microhone.

With this change, users of Superpowered will need to add SUPERPOWERED_USES_MICROPHONE=0 in the “Preprocessor Macros” section of their app’s build settings.